RABSW-995 client-initiated lustre-to-lustre data movement (#22)
Enable client-initiated lustre-to-lustre data movement.  The client will create
an NnfDataMovement resource for this instead of an RsyncNodeDataMovement
resource.  The NnfDataMovement resource will wake up the part of the
datamovement_controller that creates a PV/PVC pair for the orchestrated
filesystem and uses the CSI driver to mount that FS and the global FS into the
MPI job's pod.

The PV/PVC pair in datamovement_controller_lustre are marked so they are
reserved for each other.  The name of the NnfDataMovement is used as the name
of the storage class.

Update the .gitignore to ignore the nnf-dm daemon, so it won't be accidently
scooped up by a git-add.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>