Add Cancel to Copy Offload API
This change adds cancel functionality to the Copy Offload API. This
allows users to be able to stop a data movement that is in progress. The
resulting `State/Status` will be `Finished/Cancelled`. This is in
contrast of a successful data movement which results in
`Finished/Success`.

Example:

```
$ kubectl get rsyncnodedatamovements.dm.cray.hpe.com -A
NAMESPACE       NAME                 STATUS      STATE      AGE
rabbit-node-1   default-gfs2-7km6q   Success     Finished   79m
rabbit-node-1   default-gfs2-6mtdd   Cancelled   Finished   72m
```

The user must first cancel the data movement with a
DataMovementCancelRequest. This request sets the `Spec.Cancel` flag in
the data movement resource. The user will then get a
DataMovmentCancelResponse to indicate that the flag has been set and a
cancel has been requested. However, this **does not indicate that the
cancel has been completed**. The user must then retrieve the status of
the data movement via a DataMovementStatusRequest to check that the data
movement has been successfully cancelled.

This change also adds the `Spec.Simulate` string to the resource to
allow for testing. This allows the user to run the command specified by
`Spec.Simulate` rather than the rsync command to perform the data
movement. This is useful for running something like `sleep 60` to
simulate data movement during integration with the Copy Offload API.

Additionally, the `sharedProcessNamespace` option has been enabled for the
rsync template. This is to eliminate zombie processes that may be left
after an rsync has been cancelled.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>