Revert to Ginkgo V1.
Kubebuilder does not yet support Ginkgo v2, and sometimes craps out when we
ask it to do new things in this repo when v2 is in use.

Upgrade controller-gen from 0.6.1 to 0.9.2, which causes some annotations to
be updated in the CRDs.

Drop the use of DescribeTable for the moment.  Version 2 of datamovement is
waiting in the wings and will replace that.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>