Add Progress reporting to Data Movement (#60)
This feature adds progress reporting to data movement. The `dcp` command
used for data movement has a progress option to output the current
progress. That output is read and the progress is extracted. The
progress is then stored in the Data Movement's `CommandStatus` structure
along with some other information such as elapsed time.

The progress is collected by default every 5 seconds.

To support proper configuration and testing, the `nnf-dm-config`
config map was added to the default namespace. This allows a developer to
change the data movement command, progress collection interval, and the
number of process to supply to the mpirun command.

The config map is required for data movement. If the config map is not
found, the data movement will requeue until the config map can be read.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>