Add permissions to data movement via new dcp options
The nnf-mfu image update allows us to use the new (but not approved) dcp
`-U` and `-G` options to set the user and group ID to perform the data
movement.

Note that this change requires that the mount points for the src and
dest of the data movement also have the proper permissions. That part of
this is not yet implemented.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>

Added support for dcp user/group ID options

With an updated version of dcp that now supports user and group ID, this
change uses the uid and gid from the DM's spec and feeds it into dcp.
These options will always be used, even if uid/gid are 0.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>