Github-84: Add commit hash to RPM builds (#87)
* Github-84: Add commit hash to RPM builds

Use the GITHUB_SHA environment variable provided in the github action and save it in
the file .commit. The RPM build will use the value in this file instead of the output
provided by the git command in the Makefile. This allows the git hash to be set when
building from the source RPM.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>

* Truncate GITHUB_SHA to 8 characters

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>

---------

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>