Data Movement: Allow use of hostfiles + revamped Configuration (#89)
This allows for the configuration of slots and max_slots in the mpirun
hostfile. Coupled with this is the slimmed down configuration which 
allows for more flexibility. Users now specify the full data movement
command in the ConfigMap and use $VAR substitution.

Profiles have also been introduced in the ConfigMap. However, only 
the `default` can be used at this time.

Changelog:
* Removed the use of mpirun --host. Instead, create hostfiles and use mpirun --hostfile
* Simplified the DM command to be more flexible and use $VAR substitution
  for built-ins (i.e. hostfile name, gid/uid, src/dest).
* Changed DM ConfigMap to allow for nested values for profiles.
* ConfigMap values are now typed.
* DM ConfigMap is now part of the config/default deployment.
* Change the nnf-dm-config ConfigMap namespace from default to
  nnf-dm-system

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>