Release v0.0.2 (#102)
* Change DM Manager to use port 22 for ssh/mpirun + update nnf-mfu image (#81)

The DM manager was configured to use port 2222. However, the DM worker
nodes running sshd are only listening on port 22. To simplify things, DM
manager has gone back to port 22.

Additionally, the worker image was updated to track the latest version of
nnf-mfu.

Tested in kind VM. The manager can hit the worker via mpirun now that
it's targeting port 22:

```
$ kubectl -n nnf-dm-system exec nnf-dm-controller-manager-587495bc7-qrxhg -c manager -- mpirun --allow-run-as-root --host 10-244-3-3.dm.nnf-dm-system hostname
Warning: Permanently added '10-244-3-3.dm.nnf-dm-system,10.244.3.3' (ECDSA) to the list of known hosts.
nnf-dm-worker-8w7cs
```

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>

* Allow builds on all branches. (#82)

Loosen the branch filter for pushes.
Print some event context, for future debugging.
Remove an unused variable from verify_tag.
Rename some jobs to give them unique names, to help with debugging.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>

* New Version endpoint with support for the data-mover 'version' and a … (#83)

* New Version endpoint with support for the data-mover 'version' and a list of supported 'API versions'
* Add support for including the git commit hash when building the daemon

Signed-off-by: Nate Thornton <nate.thornton@hpe.com>

* Github-84: Add commit hash to RPM builds (#87)

* Github-84: Add commit hash to RPM builds

Use the GITHUB_SHA environment variable provided in the github action and save it in
the file .commit. The RPM build will use the value in this file instead of the output
provided by the git command in the Makefile. This allows the git hash to be set when
building from the source RPM.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>

* Truncate GITHUB_SHA to 8 characters

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>

---------

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>

* Data Movement: Allow use of hostfiles + revamped Configuration (#89)

This allows for the configuration of slots and max_slots in the mpirun
hostfile. Coupled with this is the slimmed down configuration which 
allows for more flexibility. Users now specify the full data movement
command in the ConfigMap and use $VAR substitution.

Profiles have also been introduced in the ConfigMap. However, only 
the `default` can be used at this time.

Changelog:
* Removed the use of mpirun --host. Instead, create hostfiles and use mpirun --hostfile
* Simplified the DM command to be more flexible and use $VAR substitution
  for built-ins (i.e. hostfile name, gid/uid, src/dest).
* Changed DM ConfigMap to allow for nested values for profiles.
* ConfigMap values are now typed.
* DM ConfigMap is now part of the config/default deployment.
* Change the nnf-dm-config ConfigMap namespace from default to
  nnf-dm-system

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>

* Update nnf-sos vendor

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>

* RABSW-1150: Add ServiceAccount for nnf-dm daemon (#93)

Add a ServiceAccount for clientmountd that only allows access to the resources that
the nnf-dm daemon needs.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>

* nnf-dm daemon: Increase k8s client QPS and Burst (#94)

The default values of 5 and 10 are impacting the creation of DM requests
via the Copy Offload API. 

- These values have been increased to 50 and
100. Command line parameters have been added to adjust these.
- Updated the go client to create the DM requests
concurrently and make the simulated server thread safe.
- DM Worker log output has been reduced; lots of concurrent request make the log unreadable

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>

* Use the live k8s client object in suite_test.go. (#96)

In the kubebuilder book:
https://book.kubebuilder.io/cronjob-tutorial/writing-tests.html
It explains that we should be using the "live" k8s client rather than the one
from the manager:

"Note that we set up both a “live” k8s client and a separate client from the
manager. This is because when making assertions in tests, you generally want to
assert against the live state of the API server. If you use the client from the
manager (k8sManager.GetClient), you’d end up asserting against the contents of
the cache instead, which is slower and can introduce flakiness into your
tests."

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>

* Upgrade controller-runtime and friends (#95)

Upgrade controller-runtime, ginkgo, gomega.  Revendor dws and pick up the
new API for status updater.

Upgrade controller-gen and env-k8s-version.

* Copy Offload API: Add DCPOptions, DryRun, Start/End Time; update lib-cpp (#99)

Allow an NnfDataMovement resource to define UserConfig that can be used
to override/supplement Data Movement Configuration. This way, the Copy
Offload API can tweak behavior on top of the existing DM configuration
that is defined by the nnf-dm-config ConfigMap.

This enables the Copy Offload API to:
- Add extra options to the `dcp` command
- Perform a dryrun where the data movement command becomes `/bin/true`

Changes:
- `CreateRequest` updates:
    - Added `dcpOptions`
    - Added `dryrun` (wasn't doing anything before)
- Added `startTime` and `endTime` to `StatusResponse`
- Updated API readme
- Synced lib-cpp (added CommandStatus, dryrun, dcpOptions, start/end time)
    - Added Makefiles for lib-cpp, lib-cpp/example, and client-cpp
    - Added Readme.md for lib-cpp

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>

* Use the new "lus" API group for lustre-fs-operator (#100)

Use the new "lus" API group for lustre-fs-operator

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>

* Copy Offload API: Add generated documentation (#101)

Use protoc-gen-doc to generate HTML documentation for the Copy Offload
API. This way, the code is self documenting and we do not have to update
the readme for any changes.

Updated the readme to point to the generated API documenation and to add more
clarity in certain areas.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>

* Github action triggers on master and release branches

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>

---------

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>
Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>
Signed-off-by: Nate Thornton <nate.thornton@hpe.com>
Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>
Co-authored-by: Blake Devcich <89158881+bdevcich-hpe@users.noreply.github.com>
Co-authored-by: Dean Roehrich <dean.roehrich@hpe.com>
Co-authored-by: Nate Thornton <nate.thornton@hpe.com>
Co-authored-by: Blake Devcich <blake.devcich@hpe.com>