Added slots/max_slots to Copy Offload API
slots and maxSlots can now be supplied in the Copy Offload API. This
allows the user to override the server side configuration (nnf-dm-config
ConfigMap) to set the slots/max_slots in the mpirun hostfile.

Updated nnf-sos to support this.

For copy_in/copy_out, there is equivalent at this time.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>