Added profile to Copy Offload API + fix nnf-dm logging
`DataMovementCreateRequest`can now specify a server-side DM profile via
the `profile` parameter. If empty, it will use the `default` profile.

nnf-dm was not performing `ctrl.SetLogger`, so there was no logging
happening in sysconfig. I assume this is due to the updates to
controller runtime. This has been updated to match what the DM
controllers are doing.

Additionally, fixed an issue with the API incorrectly reporting invalid
DM requests as Pending.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>