Remove the ComputeNodes array from SystemConfiguration
The list of compute node names is already found within the spec.StorageNodes
array. By having the names in just one array, rather than two, that ought to
help reduce errors.

However, the main problem was this: The spec.ComputeNodes array can account for
up to 75% of the size of the resource on a large cluster.  This has caused
problems for 'kubectl apply' (client-side-apply) when it creates the
"last-applied-configuration" annotation for the resource, with the serialized
version of the resource exceeding the maximum allowed size of an annotation.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>