Ensure destination path exists prior to data movement
To ensure that dcp behaves the same on each data movement, make sure the
destination path exists before attempting data movement.

This also ensures that as long as the user provides a valid path with the
correct permissions, dcp will have a safe place for the destination.

Path validation will need to be done upfront in nnf-sos and the nnf-dm
daemon.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>