Daemon RPM build rework
Build the daemon in an environment defined by a Dockerfile, and build it with
the same environment whether it's built locally or within a workflow.

Using `make build-daemon` will build the daemon locally as always and continue
to drop it into the local bin/ directory.

Using `make build-daemon-rpm` will build the daemon within the container
environment and will drop the resulting rpm in a local rpm/ directory.

In tests so far, this tends to cut the build time within the workflow to about
1/3 of the time that we've seen in the past.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>