Add version labels to release manifest Deployment/DaemonSet
Change make-kustomization.sh to add component/version labels only to
Deployment/Daemonset in the manifest. By doing only these resources, rather
than all resources, we minimize the noise in a manifest update to the gitops
repo.

If run locally the labels will be:

  app.kubernetes.io/component: nnf-dm
  app.kubernetes.io/version: <git-version-gen>

If run from nnf-deploy's "make manifests" the labels will be:

  app.kubernetes.io/component: nnf-dm
  app.kubernetes.io/version: <nnf-dm's git-version-gen>
  app.kubernetes.io/part-of: nnf
  app.kubernetes.io/nnf-version: <nnf-deploy's git-version-gen>

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>