Add StatCommand to NnfDataMovementProfile (#201)
The current implementation of the stat commands that are used in data
movement destination prep are hard-coded. This allows the command to be
customized in the profile.

When I first implemented this, I tried to keep the profile and dm
resource out of the function calls as much as possible and just pass
down the info needed. This change is to the contrary of that since we
now need to pass the statCommand all the way down to the bottom of the
call stack when the stat actually takes place.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>