Introduce a copy-offload daemon in a user container
This daemon is an http server that accepts copy-offload requests and
executes them directly. The daemon is in a container, on the rabbit,
and was placed there by a "#DW container" directive.

This commit includes a daemon that can be built locally and run outside
a container, or it can be built in a container image.
The necessary RBAC bits for config/ are coming in the next step.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>