Copy offload server, client, and tests (#258)
Add TLS/token support to the copy-offload server. Version the REST API.

Add a libcopyoffload library which is meant to be used by compute applications to talk to the copy-offload server. This is written in C and is built on top of libcurl.

Update the Makefile's "edit-image" and "deploy" targets to deploy the pieces for the nnf-dm-copy-offload pod. This installs template profiles for a NnfContainerProfile and a NnfDataMovementProfile.

Add a shell test that runs the copy-offload server in a mock mode, not talking to k8s. This runs the server and verifies the REST endpoints and TLS/token handling. This includes a test tool, written in C, that uses the libcopyoffload library to talk to the server.

Add tools to generate certificates and tokens that a client using libcopyoffload will use to talk to the server.

Add ginkgo tests for the http server, using httptest. This tests the REST endpoints, REST API versioning, and some token handling.

Add a github workflow to create the "nnf-dm-copy-offload" container that has the copy-offload http server.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>