Support per-Workflow tokens (#273)
Update the example copyoffload-containerprofile.yaml to no longer explicitly
mount the Secrets. This is now handled by nnf-sos when it creates the
Deployment for the user-container. Use environment variables for the path
names to the cert&key and token signing key. The environment variables will
be added to the Deployment by nnf-sos as well.

Rename tools/mk-copy-offload-secrets.sh to tools/mk-usercontainer-secrets.sh
to show that this cert is available for any user container. Also no longer
create the token secret here, because this is now handled by nnf-sos. Remove
the token generation pieces from tools/gen_certs.sh since this is now handled
by nnf-sos.

Change libcopyoffload to expect the token to be in an environment variable
named DW_WORKFLOW_TOKEN. Allow this to be overridden by specifying a file that
has the token. Update e2e-mocked.sh to use a DW_WORKFLOW_TOKEN environment
variable to pass the token to the tester tool. Add a library that e2e-mocked
can use to create the token signing key and token for its own purposes.

Change the copy-offload server to always wrangle the token signing key in
binary form from a PEM block.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>