Use $SETPRIV in mkdir/stat command when necessary
With the new copy offload container, the container is ran as the
specified user, so setpriv commands are not necessary to preserve the
proper permissions. However, for DataIn/DataOut DM, data movement is
still performed as root in the nnf-dm containers. Setpriv is necessary
in this case.

This change determines the UID of the user and appropriate uses/drops
the setprivCommand defined by the DM profile when issuing mkdir and stat
commands.

The same is done for the main DM command to remove --uid/--gid from dcp
when running in copy-offload.

These changes removes the need to have separate DM profiles for copy
offload.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>