Cleanup copy-offload library API
Simplify copy_offload_configure() by removing the token and cert paths. Let it
rely on the defaults for those. Provide two functions to override the defaults.

This adds copy_offload_override_cert() to override the default filename that
is used for the TLS certificate.

This adds copy_offload_override_token() to override the default location to
find the token.

Remove support for mTLS from the library, the server, and the e2e-mocked
tests.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>