Copy Offload API: Use MPI Operator to drive
The current copy offload API design does not work for lustre-to-lustre
data movement, where every rabbit in the workflow needs to get involved.

For background, GFS2-to-lustre data movement is node-local, meaning that
only one rabbit is used to transfer the node-local GFS2 filesystem from
the rabbit to global lustre. Since a lustre filesystem can be spread
across multiple rabbit nodes, each rabbit must be contacted to
orchestrate the data movement across the nodes.

To support this with user containers for copy offload, SSH keys must be
set up so that each rabbit can contact any other rabbit. The mpi-operator
takes care of this orchestration for us, and mpirun can be used to run
jobs (e.g., dcp) on each rabbit node.

With this model, we get one launcher pod for the workflow, and then each
rabbit gets a worker pod. The launcher pod is responsible for launching
the mpirun job across the worker pods. The launcher pod is running the Copy
Offload server, which will be contacted by the compute nodes to
communicate with the Copy Offload API.

A new environment variable is added to nnf-sos so that compute nodes
know how to contact the Copy Offload API.

For GFS2 data movement, the server must use the compute node that is
making the request and trace back to its local rabbit. The software then
must contact the worker pod that is running on that rabbit to initiate
data movement with mpirun using only that worker pod. In contrast, for
lustre data movement, all of the worker pods running on all of the
rabbits for the workflow are used.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>