Do all copy-offload request setup in driver.Create
Move all setup steps into driver.Create(). If any of that fails then the
error is reported in the HTTP result of the copy-offload request, and there
is no NnfDataMovement resource.

This means that driver.Drive() can be focused on actually running the
copy command. In the next step, any status or result from the copy command
can be recorded in the NnfDataMovement resource.

Make copy-offload server mock mode more useful by allowing to actually run
a command.

Adjust the 'cancel' test to allow commands time to terminate.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>