copy offload: Add slots/max_slots to copy_offload_copy() and tester (#290)
This was defaulted to 0, which disables slots/maxSlots on the server
side. We want this to default to -1 to defer to the slots and
maxSlots set by the DM profile.

These options allows users to set the slots/max_slots in the mpirun
hostfile for any request. This allows the user to pick how many
processes dcp should be ran with on each host (i.e. rabbit).

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>