libcopyoffload: use json-c to parse the status response (#303)
Add the json-c library to the build env. Use it in libcopyoffload to
parse the status response from the server.

Change the response for a cancel request to return StatusOK rather than
StatusNoContent.

In server_test.go, fix some tests that were incorrect but silent. Allow
some positive tests in some sections that had only negative tests. Add a
section for the status request.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>