Use NnfPodSpec in NnfDataMovementManager (#307)
NnfDataMovementManager now uses NnfPodSpec instead of a PodTemplateSpec.

The config has been updated to do this and the controller now needs to
add in things that are not supported by NnfPodSpec:
- serviceAccountName
- tolerations
- shareProcessNamespace

And similarly for NnfContainer, the worker container's security context
needs to be set.

There was an unnecessary check for NNF_TEST_ENVIRONMENT to workaround
the manager container not being present in the unit tests - that
container has been added.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>