copy-offload: Fix MPI hostfile tmpdir
Copy offload requests don't yet have a name when we start doing all the
destination prep work. This work requires mpirun operations, which in
turn requires an MPI hostfile.

The current code allows for an empty `dmName` which ends up putting the
hostfile at `/tmp/hostfile`, which is **not** unique and is then used
by multiple offload requests.

This fix address that for this case and creates a tmpdir to use.

For regular DM, keep using the name so that it can easily be tracked for
debugging purposes.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>